Gadget3 formula helpers

Tools to create R formulas

g3_formula(code, ...)

Arguments

code

Unevaluated code to be turned into a formula

...

Named items to add to the formula's environment, or a single list / environment to use.

Details

When using ~, the local environment is attached to the code. This can leak unwanted variables into a model. This allows you to avoid the problem without resorting to local.

Value

A formula object, with environment created from .... Can then be used anywhere in gadget3 that accepts a formula.

Examples

# g3_formula is identical to defining a formula within local():
stopifnot(all.equal(
    g3_formula(x + 1, z = 44),
    local({ z = 44; ~x + 1 })
    ))

# If the code is destined for CRAN, you need to quote() to avoid check errors:
stopifnot(all.equal(
    g3_formula(quote( x + 1 ), z = 44),
    local({ z = 44; ~x + 1 })
    ))

# ... we can use these in g3a_grow_lengthvbsimple(), e.g.
g3a_grow_lengthvbsimple(
    linf_f = g3_formula( g3_param("ling.Linf") ),
    kappa_f = g3_formula( g3_param("ling.k") * k, k = 0.001))
#> ~avoid_zero_vec((g3_param("ling.Linf") - stock__midlen) * (1 - 
#>     exp(-((g3_param("ling.k") * k)) * cur_step_size)))
#> <environment: 0x562b4299cee8>