Gadget3 likelihood actions

Gather nll in a g3 model

g3l_catchdistribution_sumofsquares(over = c("area"))

g3l_catchdistribution_multinomial()

g3l_catchdistribution_multivariate(rho_f, sigma_f, over = c("area"))

g3l_catchdistribution_surveyindices(mode = "log", alpha = 0, beta = 1)

g3l_catchdistribution(nll_name, obs_data, fleets, stocks, function_f, missing_val = 0,
    area_group = NULL, weight = 1, run_at = 10)

Arguments

over
When comparing proportions of lengthgroups, specifies the dimensions that define the total. For example the default "area" means the proprtion of the current lengthgroup to all individuals in that area.

rho_f
TODO:

sigma_f
TODO:

mode
Character string, either "log" or "linear".

alpha
TODO:

beta
TODO:

nll_name
Character string, used to define the variable name for obsstock and modelstock.

obs_data
Data.frame of observation data, for example the results of mfdb_sample_count

fleets
A list of g3_stock objects to collect catch data for

stocks
A list of g3_stock objects to collect catch data for

function_f
A formula to compare obsstock to modelstock and generate nll, defined by one of the g3l_catchdistribution_* functions.

missing_val
Where there are missing values in the incoming data, value to replace them with.

area_group
TODO:

weight
TODO:

run_at
Order of age actions within model, see TODO:

Details

The actions will define the following variables in your model:

obsstock__num

A g3_stock instance that contains all observations in an array

modelstock__num

A g3_stock instance that groups in an identical fashion to obsstock, that will be filled with the model's predicted values

See also

Value

g3l_catchdistribution_sumofsquares

TODO:

g3l_catchdistribution_multinomial

TODO:

g3l_catchdistribution_multivariate

TODO:

g3l_catchdistribution_surveyindices

TODO:

g3l_catchdistribution

An action (i.e. list of formula objects) that will...

For all fleets and stocks combinations, collect catch data into modelstock__num

Compare modelstock__num with obsstock__num, using function_f

Examples

library(magrittr)
ling_imm <- g3_stock('ling_imm', seq(20, 156, 4)) %>% g3s_age(3, 10)
ling_mat <- g3_stock('ling_mat', seq(20, 156, 4)) %>% g3s_age(5, 15)
lln <- g3_fleet('lln')

if (FALSE) # Fetch length distribution via. MFDB
ldist.lln <- mfdb_sample_count(mdb, c('age', 'length'), c(list(
    sampling_type = 'SEA',
    data_source = 'iceland-ldist',
    gear = c('LLN','HLN'),
    age = mfdb_interval("all",
        c(minage, maxage),
        open_ended = c("upper","lower")),
    length = mfdb_interval("len",
        seq(minlength, maxlength, by = dl),
        open_ended = c("upper","lower"))), defaults))

# Invent a ldist.lln table for our tests
ldist.lln <- list(expand.grid(
    year = 1999:2000,
    step = c(1, 2),
    age = c(5,6,7,8)))
ldist.lln[[1]]$number <- floor(runif(nrow(ldist.lln[[1]]), min=100, max=999))

likelihood_actions <- list(
  g3l_catchdistribution(
    'ldist_lln',
    ldist.lln[[1]],
    fleets = list(lln),
    stocks = list(ling_imm, ling_mat),
    g3l_catchdistribution_sumofsquares()))