Gadget3 likelihood actions

Gather nll in a g3 model

g3l_distribution_sumofsquares(over = c("area"))

g3l_distribution_multinomial(epsilon = 10)

g3l_distribution_multivariate(rho_f, sigma_f, over = c("area"))

g3l_distribution_surveyindices_log(alpha = 0, beta = 1)

g3l_distribution_surveyindices_linear(alpha = 0, beta = 1)

g3l_distribution(nll_name, obs_data, fleets = list(), stocks, function_f,
    missing_val = 0, area_group = NULL, report = FALSE, nll_breakdown = FALSE, weight = 1,
    run_at = 10)

g3l_catchdistribution(nll_name, obs_data, fleets, stocks, function_f, missing_val = 0,
    area_group = NULL, report = FALSE, nll_breakdown = FALSE, weight = 1, run_at = 10)

Arguments

over
When comparing proportions of lengthgroups, specifies the dimensions that define the total. For example the default "area" means the proprtion of the current lengthgroup to all individuals in that area.

rho_f
formula substituted into multivariate calcuations, see below.

sigma_f
formula substituted into multivariate calcuations, see below.

epsilon
Value to be used whenever the calculated probability is very unlikely. Default 10.

alpha
formula substituted into surveyindices calcuations, see below.

beta
formula substituted into surveyindices calcuations, see below.

nll_name
Character string, used to define the variable name for obsstock and modelstock.

obs_data
Data.frame of observation data, for example the results of mfdb_sample_count.

Should at least have a year column, and a length or weight column. For more information, see "obs_data and data aggregation" below.

fleets
A list of g3_stock objects to collect catch data for. If empty, will collect abundance data for stocks instead.

stocks
A list of g3_stock objects to collect catch or abundance data for, depending if stocks were provided.

function_f
A formula to compare obsstock__x to modelstock__x and generate nll, defined by one of the g3l_distribution_* functions.

This will be adapted to compare either number (modelstock__num) or weight (modelstock__wgt) depending on what columns obs_data has.

missing_val
Where there are missing values in the incoming data, value to replace them with.

area_group
mfdb_group or list mapping area names used in obs_data to integer model areas, see "obs_data and data aggregation" below.

report
If TRUE, add model and observation arrays to the model report, called cdist_nll_name_model__num/wgt and cdist_nll_name_obs__num/wgt respectively

nll_breakdown
Should the nll report be broken down by time? TRUE / FALSE

weight
Weighting applied to this likelihood component.

run_at
Integer order that actions will be run within model, see "Ordering of actions" in vignette("model_structure").

Details

The actions will define the following variables in your model:

obsstock__num/wgt

A g3_stock instance that contains all observations in an array

modelstock__num/wgt

A g3_stock instance that groups in an identical fashion to obsstock, that will be filled with the model's predicted values

The model report will contain nll_cdist_nll_name__num and/or nll_cdist_nll_name__wgt, depending on the columns in obs_data (a number column will compare by individuals, and produce a corresponding num report). If nll_breakdown is TRUE, this will be an array with one entry per timestep.

The only difference between g3l_distribution and g3l_catchdistribution is the latter insists that at least one fleet is provided, so it can compare catch rather than abundance.

obs_data and data aggregation

The obs_data data.frame, as well as providing the observation data to compare the model data against, controls the grouping of model data to compare to the observation data, by inspecting the MFDB column attributes produced by e.g. mfdb_sample_count.

Metadata columns describe the observation datapoint in that row. The columns should be from this list:

year

Required. Year for the data point. Gaps in years will result in no comparison for that year

step

Optional. If there is no step column, then the data is assumed to be yearly, and the model data for all timesteps will be summed before comparing. Model timestep for the data point. Gaps in steps will result in no comparison for that year/step.

length

Optional. If missing all lengthgroups in the model will be summed to compare to the data. If open_ended = c('lower', 'upper') was used when querying MFDB for the data, then the bottom/top length groups will be modified to start from zero or be infinite respectively. Any missing lengthgroups (when there is otherwise data for that year/step) will be compared to zero.

age

Optional. If missing all age-groups (if any) in the model will be summed to compare to the data. Model ages will be grouped by the same groupings as MFDB used, thus if the data was formed with a query age = mfdb_group(young = 1:3, old = 4:5), then the model data will similarly have 2 groups in it. Any missing ages (when there is otherwise data for that year/step) will be compared to zero.

stock

Optional. If missing all stocks in stocks will be summed to compare to the data. The values in the stocks column should match the names of the stocks given in the stocks parameter. This column can be factor or character. Any missing stocks (when there is otherwise data for that year/step) will be compared to zero.

area

Optional. If missing all areas in the model will be summed to compare to the data. Unlike other columns, the MFDB grouping here is ignored (the areas it is grouping over aren't integer model areas). Instead, the area_group parameter should describe how to map from the area names used in the table to integer model areas. For example, if area_group = list(north=1:2, south=3:5), then the area column of obs_data should contain either "north" or "south", and corresponding model data will be summed from integer model areas 1,2 and 3,4,5 respectively. If area_group is not supplied, then we assume that obs_data area column will contain model area integers. Any missing areas (when there is otherwise data for that year/step) will be compared to zero.

Data columns contain the observation data to compare. There should be at least one of:

number

If a number column appears in obs_data, then the stock abundance by individuals will be aggregated and compared to the obs_data number column.

weight

If a weight column appears in obs_data, then the total biomass of the stock will be aggregated and compared to the obs_data number column.

See also

Value

g3l_distribution_sumofsquares

Returns a formula for use as function_f:

$$ \sum_{\it lengths} \Big( \frac{N_{tral}}{N_{tr}} - \frac{\nu_{tral}}{\nu_{tr}} \Big) ^2 $$

\(N_{tral}\)

Observation sample size for current time/area/age/length combination

\(\nu_{tral}\)

Model sample size for current time/area/age/length combination

\(N_{tr}\)

Total observation sample size for current time/area (or dimensions set in over)

\(\nu_{tr}\)

Total model sample size for current time/area (or dimensions set in over)

g3l_distribution_multinomial

Returns a formula for use as function_f:

$$ 2 ( \sum_{\it lengths} \log N_{tral}! - \log (\sum_{\it lengths} N_{tral})! - \sum_{\it lengths} ( N_{tral} \log min(\frac{\nu_{tral}}{\sum_{\it lengths} \nu_{tral}}, \frac{1}{l \epsilon}) ) ) $$

\(N_{tral}\)

Observation sample size for current time/area/age/length combination

\(\nu_{tral}\)

Model sample size for current time/area/age/length combination

\(l\)

Number of lengthgroups in sample

\(\epsilon\)

epsilon parameter

g3l_distribution_multivariate

Returns a formula for use as function_f, which calls TMB's SCALE(AR1(rho), sigma)(x), where rho and sigma are parameters, and x is defined as:

$$ \frac{N_{tral}}{N_{tr}} - \frac{\nu_{tral}}{\nu_{tr}} $$

\(N_{tral}\)

Observation sample size for current time/area/age/length combination

\(\nu_{tral}\)

Model sample size for current time/area/age/length combination

\(N_{tr}\)

Total observation sample size for current time/area (or dimensions set in over)

\(\nu_{tr}\)

Total model sample size for current time/area (or dimensions set in over)

For more information, see http://kaskr.github.io/adcomp/_book/Densities.html#autoregressive-processes in the TMB book.

g3l_distribution_surveyindices_log

Returns a formula for use as function_f:

$$ \sum_{\it lengths} (\alpha + \beta \log N_{tral} - \log \nu_{tral})^2 $$

\(N_{tral}\)

Observation sample size for current time/area/age/length combination

\(\nu_{tral}\)

Model sample size for current time/area/age/length combination

\(\alpha\)

alpha parameter

\(\beta\)

beta parameter

g3l_distribution_surveyindices_linear

Returns a formula for use as function_f:

$$ \sum_{\it lengths} (\alpha + \beta N_{tral} - \nu_{tral})^2 $$

\(N_{tral}\)

Observation sample size for current time/area/age/length combination

\(\nu_{tral}\)

Model sample size for current time/area/age/length combination

\(\alpha\)

alpha parameter

\(\beta\)

beta parameter

g3l_distribution

Assuming fleets is empty, an action (i.e. list of formula objects) that will...

For all stocks, collect catch data into modelstock__num or modelstock__wgt, depending on the columns provided in obs_data

Compare modelstock__num/wgt with obsstock__num/wgt, using function_f

The output of function_f is summed over all stock dimensions (age/area) and time and added to nll.

g3l_catchdistribution

An action (i.e. list of formula objects) that will...

For all fleets and stocks combinations, collect catch data into modelstock__num or modelstock__wgt, depending on the columns provided in obs_data

Compare modelstock__num/wgt with obsstock__num/wgt, using function_f

The output of function_f is summed over all stock dimensions (age/area) and time and added to nll.

Examples

library(magrittr)
ling_imm <- g3_stock('ling_imm', seq(20, 156, 4)) %>% g3s_age(3, 10)
ling_mat <- g3_stock('ling_mat', seq(20, 156, 4)) %>% g3s_age(5, 15)
lln <- g3_fleet('lln')

if (FALSE) # Fetch length distribution via. MFDB
ldist.lln <- mfdb_sample_count(mdb, c('age', 'length'), c(list(
    sampling_type = 'SEA',
    data_source = 'iceland-ldist',
    gear = c('LLN','HLN'),
    age = mfdb_interval("all",
        c(minage, maxage),
        open_ended = c("upper","lower")),
    length = mfdb_interval("len",
        seq(minlength, maxlength, by = dl),
        open_ended = c("upper","lower"))), defaults))

# Invent a ldist.lln table for our tests
ldist.lln <- list(expand.grid(
    year = 1999:2000,
    step = c(1, 2),
    age = c(5,6,7,8)))
ldist.lln[[1]]$number <- floor(runif(nrow(ldist.lln[[1]]), min=100, max=999))

likelihood_actions <- list(
  g3l_catchdistribution(
    'ldist_lln',
    ldist.lln[[1]],
    fleets = list(lln),
    stocks = list(ling_imm, ling_mat),
    g3l_distribution_sumofsquares()))