Gadget3 CKMR likelihood

*Experimental* CKMR tagging likelihood

Usage

g3l_tagging_ckmr(
        nll_name,
        obs_data,
        fleets,
        parent_stocks,
        offspring_stocks,
        weight = substitute(
            g3_param(n, optimise = FALSE, value = 1),
            list(n = paste0(nll_name, "_weight"))),
        run_at = g3_action_order$likelihood)

Arguments

nll_name

Character string, used to define the variable name for obsstock and modelstock.

obs_data

Data.frame of observed mother-offspring pairs with columns year / parent_age / offspring_age / mo_pairs

fleets

A list of g3_stock objects to collect catch data for.

parent_stocks

A list of g3_stock objects that are parents in a g3a_spawn action

offspring_stocks

A list of g3_stock objects that are output_stocks in a g3a_spawn action

weight

Weighting applied to this likelihood component. Default is a g3_param that defaults to 1, allowing weights to be altered without recompiling.

run_at

Integer order that actions will be run within model, see g3_action_order.

Details

Implementation of CKMR based on Bravington, M.V., Skaug, H.J., & Anderson, E.C. (2016). Close-Kin Mark-Recapture. Statistical Science, 31, 259-274.

Only one kinship probability is implemented, mother-offspring with lethal sampling, i.e. (3.2) in the paper. This is then used as a pseudo-likelihood as per (4.1).

obs_data

The obs_data data.frame provides observed pairs. Unlike other likelihood mehthods, it has a fixed structure:

year

Year of observation for the data point.

parent_age

Age of the parent in an observed parent-offspring pair.

offspring_age

Age of the offspring in an observed parent-offspring pair.

mo_pairs

Number of pairs observed with these ages.

See also

Value

g3l_tagging_ckmr

An action (i.e. list of formula objects) that will...

For all parent_stocks and offspring_stocks, collect spawing rate into modelhist__spawning and modelhist__spawned, total number of parents and total number of spawned offspring respectively

For all fleets, collect catch data into modelhist__catch

For any observed pairs that year, include the probability of that event happening into nll