Gadget3 suitability formulae

Formula-returning functions describing predator/prey suitability relationships.

g3_suitability_exponentiall50(alpha, l50)

g3_suitability_andersen(p0, p1, p2, p3 = p4, p4, p5 = ~prey_stock__midlen)

g3_suitability_gamma(alpha, beta, gamma)

g3_suitability_exponential(alpha, beta, gamma, delta)

g3_suitability_straightline(alpha, beta)

g3_suitability_constant(alpha)

g3_suitability_richards(p0, p1, p2, p3, p4)

Arguments

alpha
formula substituted into calcuations, see below.

l50
formula substituted into calcuations, see below.

p0
formula substituted into calcuations, see below.

p1
formula substituted into calcuations, see below.

p2
formula substituted into calcuations, see below.

p3
formula substituted into calcuations, see below.

p4
formula substituted into calcuations, see below.

p5
formula substituted into calcuations, see below.

beta
formula substituted into calcuations, see below.

gamma
formula substituted into calcuations, see below.

delta
formula substituted into calcuations, see below.

Value

g3_suitability_exponentiall50

Returns a formula for use in predation function's suitabilities argument:

$$ \frac{1}{ 1 + e^{-\alpha (l - l_{50})} } $$

\(l\)

Vector of prey midlength for each lengthgroup

\(l_{50}\)

Length of the prey with a 50% probability of predation, from parameter l50

g3_suitability_andersen

Returns a formula for use in predation function's suitabilities argument:

$$ p_0 + p_2 e^{-\frac{(\log\frac{p_5}{l} - p_1)^2}{p_4}} $$

\(L\)

Vector of predator midlength for each lengthgroup

\(l\)

Vector of prey midlength for each lengthgroup

\(p_0\) .. \(p_2\)

Function parameter p0 .. p2

\(p_3\)

Currently unsupported. Should be used in place of \(p_3\) if \(\log\frac{L}{l} <= p_1\)

\(p_4\)

Function parameter p4

\(p_5\)

Function parameter p5, if unspecified uses \(L\), Vector of predator midlength for each lengthgroup.

NB: Specifying \(p_5\) is equivalent to using the andersenfleet function in gadget2.

g3_suitability_gamma

Returns a formula for use in predation function's suitabilities argument:

$$ (\frac{l}{(\alpha - 1)\beta\gamma}) ^ {(\alpha - 1) e^{\alpha - 1 - \frac{l}{\beta\gamma}}} $$

\(l\)

Vector of prey midlength for each lengthgroup

\(\alpha\)

Function parameter alpha

\(\beta\)

Function parameter beta

\(\gamma\)

Function parameter gamma

This is a suitability function that is more suitable for use when considering the predation by a fleet, where the parameter \(\gamma\) would represent the size of the mesh used by the fleet (specified in centimetres).

g3_suitability_exponential

Returns a formula for use in predation function's suitabilities argument:

$$ \frac{\delta}{1 + e^{-\alpha - \beta l - \gamma L}} $$

\(L\)

Vector of predator midlength for each lengthgroup

\(l\)

Vector of prey midlength for each lengthgroup

\(\alpha\)

Function parameter alpha

\(\beta\)

Function parameter beta

\(\gamma\)

Function parameter gamma

\(\delta\)

Function parameter delta

g3_suitability_straightline

Returns a formula for use in predation function's suitabilities argument:

$$ \alpha + \beta l $$

\(l\)

Vector of prey midlength for each lengthgroup

\(\alpha\)

Function parameter alpha

\(\beta\)

Function parameter beta

g3_suitability_constant

Returns a formula for use in predation function's suitabilities argument:

$$ \alpha $$

\(\alpha\)

Function parameter alpha

g3_suitability_richards

Returns a formula for use in predation function's suitabilities argument:

$$ {\big( \frac{p_3}{1 + e^{-p_0 - p_1 l - p_2 L}} \big)}^{\frac{1}{p_4}} $$

\(L\)

Vector of predator midlength for each lengthgroup

\(l\)

Vector of prey midlength for each lengthgroup

\(p_0\) .. \(p_4\)

Function parameter p0 .. p4

This is an extension to g3_suitability_exponential.

Examples

library(magrittr)
ling_imm <- g3_stock('ling_imm', seq(20, 156, 4)) %>% g3s_age(3, 10)
ling_mat <- g3_stock('ling_mat', seq(20, 156, 4)) %>% g3s_age(5, 15)
igfs <- g3_fleet('igfs')

igfs_landings <-
  structure(expand.grid(year=1990:1994, step=2, area=1, total_weight=1),
            area_group = list(`1` = 1))

# Generate a fleet predation action using g3_suitability_exponentiall50
predate_action <- g3a_predate_fleet(
    igfs,
    list(ling_imm, ling_mat),
    suitabilities = list(
        ling_imm = g3_suitability_exponentiall50(
            ~g3_param('ling.igfs.alpha'), ~g3_param('ling.igfs.l50')),
        ling_mat = g3_suitability_exponentiall50(
            ~g3_param('ling.igfs.alpha'), ~g3_param('ling.igfs.l50'))),
    catchability = g3a_predate_catchability_totalfleet(
        g3_timeareadata('igfs_landings', igfs_landings)))