Gadget3 growth action

Add growth/maturity actions to a g3 model

g3a_grow_lengthvbsimple(linf_f, kappa_f)

g3a_grow_weightsimple(alpha_f, beta_f)

g3a_grow_impl_bbinom(delta_len_f, delta_wgt_f, beta_f, maxlengthgroupgrowth)

g3a_growmature(stock, impl_f, maturity_f = ~0, output_stocks = list(), 
    output_ratios = rep(1/length(output_stocks), times = length(output_stocks)), 
    transition_f = ~cur_step_final, run_f = ~TRUE, run_at = 5, 
    transition_at = 7)

Arguments

linf_f
A formula to substitute for \(L_\infty\).

kappa_f
A formula to substitute for \(\kappa\).

alpha_f
A formula to substitute for \(\alpha\).

beta_f
A formula to substitute for \(\beta\).

maxlengthgroupgrowth
An integer with the maximum length groups an individual can jump in one step.

stock
g3_stock to grow.

delta_len_f
A formula defining a non-negative vector for mean increase in length for stock for each lengthgroup, as defined by g3a_grow_lengthvbsimple.

delta_wgt_f
A formula defining the corresponding weight increase as a matrix of lengthgroup to lengthgroup delta for stock, as defined by g3a_grow_weightsimple.

impl_f
A pair of formula objects, as defined by g3a_grow_impl_bbinom. Both define a matrix of length groups i to length group deltas j (0..maxlengthgroupgrowth), the values in the first indicate the proportion of individuals moving from i to i + j, the values in the second indicate the corresponding weight increase of individuals moving from i to i + j.

maturity_f
A maturity formula, as defined by g3a_mature_constant.

output_stocks
List of g3_stocks that maturing stock should move into.

output_ratios
Vector of proportions for how to distribute into output_stocks, summing to 1, default evenly spread.

transition_f
formula specifying a contition for running maturation steps as well as growth, default final step of year.

run_f
formula specifying a condition for running this action, default always runs.

run_at
Integer order that actions will be run within model, see "Ordering of actions" in vignette("model_structure").

transition_at
Integer order that transition actions will be run within model, see "Ordering of actions" in vignette("model_structure").

Details

A model can have any number of g3a_growmature actions, so long as the calling arguments are different. For instance, run_f = ~age == 5 and run_f = ~age == 7.

impl_f's dependent variables are analysed to see what will affect growth. If nothing but cur_step_size will affect growth, then growth will only be recalculated when the step size changes.

Value

g3a_grow_lengthvbsimple

Returns a formula for use as delta_len_f:

$$ {{\Delta}L}_i = ( L_\infty - L_i )(1 - e^{-\kappa{\Delta}t}) $$

Where \( \Delta{t} \) is the length of the current timestep.

g3a_grow_weightsimple

Returns a formula for use as delta_wgt_f:

$$ {{\Delta}W}_{i,j} = \alpha ( (L_i + {{\Delta}L}_j)^\beta - {L_i}^\beta ) $$

Where \( \Delta{t} \) is the length of the current timestep, \( {{\Delta}L} \) is all possible length group increases i.e 0..maxlengthgroupgrowth.

g3a_grow_impl_bbinom

formula object converting mean growths using beta-binomia distribution. See https://gadget-framework.github.io/gadget2/userguide/chap-stock.html#beta-binomial

g3a_growmature

An action (i.e. list of formula objects) that will, for the given stock...

Move any maturing individuals into temporary storage, stock__transitioning_num / stock__transitioning_wgt

Calculate increase in length/weight using growth_f and impl_f

Move the contents of the temporary storage into output_stocks

See also

Examples

ling_imm <- g3_stock('ling_imm', seq(20, 156, 4))
ling_mat <- g3_stock('ling_mat', seq(20, 156, 4))

# Growth / maturity for immature ling
growth_action <- g3a_growmature(ling_imm,
    impl_f = g3a_grow_impl_bbinom(
        g3a_grow_lengthvbsimple(
            linf_f = ~g3_param("ling.Linf"),
            kappa_f = ~g3_param("ling.k") * 0.001),
        g3a_grow_weightsimple(
            alpha_f = ~g3_param("lingimm.walpha"),
            beta_f = ~g3_param("lingimm.wbeta")),
        beta_f = ~g3_param("ling.bbin") * 10,
        maxlengthgroupgrowth = 15),
    maturity_f = g3a_mature_constant(
        alpha = ~0.001 * g3_param("ling.mat1"),
        l50 = ~g3_param("ling.mat2")),
        output_stocks = list(ling_mat))