Gadget3 natural mortality action

Add natural mortality to a g3 model

g3a_naturalmortality_exp(param_f)

g3a_naturalmortality(stock, mortality_f, run_f = TRUE, run_at = 4)

Arguments

param_f
A formula to substitute for \(m\).

stock
g3_stock mortality applies to.

mortality_f
A mortality formula, as defined by g3a_naturalmortality_exp.

run_f
formula specifying a condition for running this action, default always runs.

run_at
Order of age actions within model, see TODO:

Value

g3a_naturalmortality_exp

A formula object with the following equation

$$ e^{-m \Delta t} $$

\(\Delta t\)

length of current timestep

g3a_naturalmortality

An action (i.e. list of formula objects) that will, for the given stock...

Remove a proportion of each stock group as calculated by the mortality formula mortality_f

See also

Examples

library(magrittr)
ling_imm <- g3_stock('ling_imm', seq(20, 156, 4)) %>% g3s_age(3, 10)

# Natural mortality for immature ling
naturalmortality_action <- g3a_naturalmortality(ling_imm,
    g3a_naturalmortality_exp(~g3_param_vector("lingimm.M")[[age - 3 + 1]]))