Gadget3 predation actions

Add predation to a g3 model

g3a_predate_catchability_totalfleet(E)

g3a_predate_catchability_numberfleet(E)

g3a_predate_catchability_linearfleet(E)

g3a_predate_catchability_effortfleet(catchability_fs, E)

g3a_predate_catchability_quotafleet(quota_table, E,
                                    sum_stocks = list(),
                                    recalc_f = NULL)

g3a_predate_fleet(fleet_stock, prey_stocks, suitabilities, catchability_f,
    overconsumption_f = ~0.96 - logspace_add_vec((0.96 - stock__consratio) * 100
    , 0.96)/100, run_f = ~TRUE, run_at = 3)

# NB: Deprecated interface, use g3a_predate_fleet with g3a_predate_catchability_totalfleet
g3a_predate_totalfleet(fleet_stock, prey_stocks, suitabilities, amount_f,
    overconsumption_f = ~0.96 - logspace_add_vec((0.96 - stock__consratio) * 100
    , 0.96)/100, run_f = ~TRUE, run_at = 3)

Arguments

fleet_stock
g3_stock that describes the harvesting fleet.

prey_stocks
List of g3_stocks that maturing stock should move into.

suitabilities
List of formula objects defining suitability of a given stock group, as defined by g3_suitability_exponentiall50.

catchability_f
A formula generated by e.g. g3a_predate_catchability_totalfleet expressing the catchability (i.e. total biomass) for that fleet.

E
A formula that returns the total biomass a stock can harvest in the current time/area, generally defined by a g3_timeareadata table.

catchability_fs
A list of formula objects, one per stock.

quota_table
A data.frame with 'biomass' and 'quota' columns, where 'biomass' corresponds to total biomass of sum_stocks stocks, the final value always being Inf. A quota is chosen by the largest 'biomass' value less than the current total biomass amount. For example, data.frame(biomass = c(100, Inf), quota = I(list(~g3_param('quota.low'), ~g3_param('quota.high'))))

sum_stocks
Either a list of g3_stock objects to sum when choosing a value from quote_table, or NULL, in which case choose the quota based on the current prey.

recalc_f
A formula denoting when to recalculate the current quota. For example ~cur_step == 1 will ensure the quota is only recalculated at the beginning of the year.

amount_f
Equivalent to E passed to g3a_predate_catchability_totalfleet.

overconsumption_f
Overconsumption rule, a formula that should cap all values in stock__consratio to <= 95 You can emulate gadget2 more closely than the default with the following overconsumption_f = quote(pmin(stock__consratio, 0.95))) (NB: Not differentiable!)

run_f
formula specifying a condition for running this action, default always runs.

run_at
Integer order that actions will be run within model, see "Ordering of actions" in vignette("model_structure").

Details

The actions will define the following stock instance variables for each given fleet_stock and prey_stock:

prey_stock__suit_fleet_stock

Suitability of (prey_stock) for (fleet_stock), in a prey array. i.e. the result of calculating the formula in suitabilities for the current state of the prey

prey_stock__fleet_stock

Biomass of (prey_stock) caught by (fleet_stock), in a prey array

fleet_stock__catch

Biomass of all prey caught by (fleet stock), in a fleet array

prey_stock__totalpredate

Biomass of total consumed (prey_stock), in a prey array

prey_stock__consratio

Ratio of prey_stock__totalpredate / (current biomass), capped by overconsumption_f

A model can have any number of g3a_predate_* actions, so long as the calling arguments are different. For instance, run_f = ~age == 5 and run_f = ~age == 7.

Value

g3a_predate_catchability_totalfleet

A formula object: $$ \frac{E N_{sl} W_{sl}}{\displaystyle \sum_{\it stocks} \sum_{\it lengths} N_{sl} W_{sl}} $$

\(E\)

E argument, biomass caught by fleet

\(N\)

Number of stock in length cell

\(W\)

Mean weight of stock in length cell

g3a_predate_catchability_numberfleet

A formula object: $$ \frac{E N_{sl}}{\displaystyle \sum_{\it stocks} \sum_{\it lengths} N_{sl}} $$

\(E\)

E argument, numbers caught by fleet

\(N\)

Number of stock in length cell

g3a_predate_catchability_linearfleet

A formula object: $$ E \Delta t N_{sl} W_{sl} $$

\(E\)

E argument, numbers caught by fleet

\(\Delta t\)

length of current timestep

\(N\)

Number of stock in length cell

\(W\)

Mean weight of stock in length cell

g3a_predate_catchability_effortfleet

A formula object: $$ E N_{sl} W_{sl} c_{s} $$

\(E\)

E argument, numbers caught by fleet

\(c_{s}\)

catchability_fs argument for the current stock

\(N\)

Number of stock in length cell

\(W\)

Mean weight of stock in length cell

g3a_predate_catchability_quotafleet

A formula object: $$ q E \Delta t N_{sl} W_{sl} $$

\(q\)

quota chosen from quota_table, corresponding to the total biomass of sum_stocks

\(E\)

E argument, numbers caught by fleet

\(\Delta t\)

length of current timestep

\(N\)

Number of stock in length cell

\(W\)

Mean weight of stock in length cell

...if recalc_f is set, this will only be recaculated when true. Any other step will use the previous value.

g3a_predate_fleet

An action (i.e. list of formula objects) that will...

Zero fleet and prey catch counters

For each prey, collect maximum desired by fleet for all prey, into a prey_stock__fleet_stock variable

After all fleet consumption is done, scale consumption using catchability_f, sum into prey_stock__totalpredate

After all consumption is done, temporarily convert prey_stock__fleet_stock to a proprotion of prey_stock__totalpredate

Calculate prey_stock__consratio (ratio of consumed to available), capping using overconsumption_f. Update prey_stock__num

Recalculate prey_stock__fleet_stock, fleet_stock__catch, post-overconsumption

See also

Examples

library(magrittr)
areas <- c(a = 1, b = 2)
ling_imm <- g3_stock('ling_imm', seq(20, 156, 4)) %>% g3s_livesonareas(areas[c('a')])
ling_mat <- g3_stock('ling_mat', seq(20, 156, 4)) %>% g3s_livesonareas(areas[c('b')])
lln <- g3_fleet('lln') %>% g3s_livesonareas(areas[c('a', 'b')])

if (FALSE) # Fetch landings data via. MFDB
lln_landings <- mfdb_sample_totalweight(mdb, NULL, c(list(
  gear=c('HLN','LLN'),
  sampling_type = 'LND',
  species = defaults$species),
  defaults))

# Invent a lln_landings table
lln_landings <- list(expand.grid(
    year = 1999:2000,
    step = c(1, 2),
    area = areas[c('a', 'b')]))
lln_landings[[1]]$total_weight <- floor(runif(nrow(lln_landings[[1]]), min=100, max=999))

predate_action <- g3a_predate_fleet(
    lln,
    list(ling_imm, ling_mat),
    suitabilities = list(
        ling_imm = g3_suitability_exponentiall50(
            ~g3_param('ling.lln.alpha'),
            ~g3_param('ling.lln.l50')),
        ling_mat = g3_suitability_exponentiall50(
            ~g3_param('ling.lln.alpha'),
            ~g3_param('ling.lln.l50'))),
    catchability_f = g3a_predate_catchability_totalfleet(
        g3_timeareadata('lln_landings', lln_landings[[1]])))