Gadget3 predation actions

Add predation to a g3 model

g3a_predate_totalfleet(fleet_stock, prey_stocks, suitabilities, amount_f, run_at = 3)

Arguments

fleet_stock
g3_stock that describes the harvesting fleet.

prey_stocks
List of g3_stocks that maturing stock should move into.

suitabilities
List of formula objects defining suitability of a given stock group, as defined by g3_suitability_exponentiall50.

amount_f
A formula that returns the total biomass a stock can harvest in the current time/area, generally defined by a g3_timeareadata table.

run_at
Order of age actions within model, see TODO:

Details

The actions will define the following stock instance variables for each given fleet_stock and prey_stock:

prey_stock__fleet_stock

Total biomass selected for that fleet/prey combination

fleet_stock__catch

Total biomass caught by that fleet

prey_stock__totalpredate

Total biomass removed from that stock

prey_stock__overconsumption

Factor to scale consumption down to at most 95% of available stock

A model can have any number of g3a_predate_* actions, so long as the calling arguments are different. For instance, run_f = ~age == 5 and run_f = ~age == 7.

Value

g3a_predate_totalfleet

An action (i.e. list of formula objects) that will...

Zero fleet and prey catch counters

For each prey, collect maximum desired by fleet for all prey, into a prey_stock__fleet_stock variable

After all fleet consumption is done, scale consumption using amount_f, sum into prey_stock__totalpredate

After all other consumption is done, work out an overconsumption co-efficient for each prey

For prey_stock__fleet_stock, fleet_stock__catch, scale by prey's overconsumption co-efficient

See also

Examples

library(magrittr)
areas <- c(a = 1, b = 2)
ling_imm <- g3_stock('ling_imm', seq(20, 156, 4)) %>% g3s_livesonareas(areas[c('a')])
ling_mat <- g3_stock('ling_mat', seq(20, 156, 4)) %>% g3s_livesonareas(areas[c('b')])
lln <- g3_fleet('lln') %>% g3s_livesonareas(areas[c('a', 'b')])

if (FALSE) # Fetch landings data via. MFDB
lln_landings <- mfdb_sample_totalweight(mdb, NULL, c(list(
  gear=c('HLN','LLN'),
  sampling_type = 'LND',
  species = defaults$species),
  defaults))

# Invent a lln_landings table
lln_landings <- list(expand.grid(
    year = 1999:2000,
    step = c(1, 2),
    area = areas[c('a', 'b')]))
lln_landings[[1]]$total_weight <- floor(runif(nrow(lln_landings[[1]]), min=100, max=999))

predate_action <- g3a_predate_totalfleet(
    lln,
    list(ling_imm, ling_mat),
    suitabilities = list(
        ling_imm = g3_suitability_exponentiall50(
            ~g3_param('ling.lln.alpha'),
            ~g3_param('ling.lln.l50')),
        ling_mat = g3_suitability_exponentiall50(
            ~g3_param('ling.lln.alpha'),
            ~g3_param('ling.lln.l50'))),
    amount_f = g3_timeareadata('lln_landings', lln_landings[[1]]))