Gadget3 renewal actions

Add renewal / initialconditions to a g3 model

g3a_initialconditions(stock, num_f, wgt_f, run_f = ~cur_time == 0L, run_at = 0)

g3a_initialconditions_normalparam(stock, factor_f, mean_f, stddev_f, alpha_f, beta_f, 
    run_f = ~cur_time == 0L, run_at = 0)

g3a_renewal(stock, num_f, wgt_f, run_f = ~TRUE, run_at = 8)

g3a_renewal_normalparam(stock, factor_f, mean_f, stddev_f, alpha_f, beta_f, 
    run_f = ~TRUE, run_at = 8)

Arguments

stock
The g3_stock to apply to

num_f
formula that produces a lengthgroup vector of number of individuals for the current age/area/... length group.

wgt_f
formula that produces a lenghgroup vector of mean weight for the current age/area/... length group.

run_at
Integer order that actions will be run within model, see "Ordering of actions" in vignette("model_structure").

factor_f
formula substituted into normalparam calcuations, see below.

mean_f
formula substituted into normalparam calcuations, see below.

stddev_f
formula substituted into normalparam calcuations, see below.

alpha_f
formula substituted into normalparam calcuations, see below.

beta_f
formula substituted into normalparam calcuations, see below.

run_f
formula specifying a condition for running this action, default always runs for renewal, first timestep for initialcondtions.

Details

A model can have any number of g3a_renewal_* actions, so long as the calling arguments are different. For instance, run_f = ~age == 5 and run_f = ~age == 7.

The g3a_renewal_* actions will define the following stock instance variables for stock:

stock__renewalnum

Extra individuals added to the stock

stock__renewalwgt

Mean weight of added individuals

Value

g3a_initialconditions / g3a_renewal

An action (i.e. list of formula objects) that will, for the given stock, iterate over each area/age/etc. combination, and generate a lengthgroup vector of new individuals and weights using num_f and wgt_f.

renewal will add fish to the existing collection, whereas initialconditions will assume the stock is currently empty.

g3a_initialconditions_normalparam / g3a_renewal_normalparam

As g3a_initialconditions / g3a_renewal, but the following formulae are used to calculate num/wgt:

$$n = e^{-(\frac{L - \mu}{\sigma})^2 / 2}$$ $$N = F 10000 \frac{n}{\sum n}$$ $$W = \alpha L^{\beta}$$

\(L\)

Midlength of length groups for current area/age/...

\(F\)

Substituted for factor_f

\(\mu\)

Substituted for mean_f

\(\sigma\)

Substituted for stddev_f

\(\alpha\)

Substituted for alpha_f

\(\beta\)

Substituted for beta_f

See also

Examples

library(magrittr)
ling_imm <- g3_stock('ling_imm', seq(20, 156, 4)) %>% g3s_age(3, 10)

initialconditions_action <- g3a_initialconditions_normalparam(
    ling_imm,
    # NB: area & age factor together (gadget2 just multiplied them)
    factor_f = ~g3_param("lingimm.init.scalar") *
        exp(-1 * (g3_param_vector("lingimm.M")[[age - 3 + 1]] + g3_param("ling.init.F")) * age) *
        g3_param_vector("lingimm.init")[[age - 3 + 1]],
    mean_f = ~g3_param("ling.Linf") * (1 - exp(-1 *
        (0.001 * g3_param("ling.k")) *
        (age - (1 + log(1 - g3_param("ling.recl") / g3_param("ling.Linf")) /
            (0.001 * g3_param("ling.k")))))),
    stddev_f = ~g3_param_vector("ling.init.sd")[[age]],
    alpha_f = ~g3_param("lingimm.walpha"),
    beta_f = ~g3_param("lingimm.wbeta"))
renewal_action <- g3a_renewal_normalparam(
    ling_imm,
    factor_f = ~g3_param("ling.rec.scalar") *
        g3_param_vector("ling.rec")[[cur_year - start_year + 1]],
    mean_f = ~g3_param("ling.Linf") * (1 - exp(-1 *
        (0.001 * g3_param("ling.k")) *
        (age - (1 + log(1 - g3_param("ling.recl") / g3_param("ling.Linf")) /
            (0.001 * g3_param("ling.k")))))),
    stddev_f = ~g3_param_vector('ling.init.sd')[[age]],
    alpha_f = ~g3_param("lingimm.walpha"),
    beta_f = ~g3_param("lingimm.wbeta"),
    run_f = ~cur_step == 1 && age == 3)