Gadget3 report actions

Add report to a g3 model

g3a_report_stock(report_stock, input_stock, instance_name, run_f = TRUE,
    run_at = 11)

Arguments

report_stock
The g3_stock to aggregate into

input_stock
The g3_stock that will be aggregated

instance_name
The g3_stock instance variable to aggregate, e.g. "num", "wgt", "catch", ...

run_f
formula specifying a condition for running this action, default always runs.

run_at
Order of age actions within model, see TODO:

Details

The actions will define the following variables in your model:

report_stock__instance_name

Results of collating input_stock__instance_name

Value

g3a_report_stock

An action (i.e. list of formula objects) that will...

Iterate over input_stock, collecting data into report_stock

Add the contents of report_stock__instance_name to the model report

See also

Examples

library(magrittr)
ling_imm <- g3_stock('ling_imm', seq(20, 156, 4)) %>% g3s_age(3, 10)

# Report that aggregates ages together
agg_report <- g3_stock('agg_report', c(1)) %>%
    g3s_agegroup(list(young = 1:3, old = 4:5)) %>%
    g3s_time(year = 2000:2002)
# Generate dissaggregated report by cloning the source stock, adding time
raw_report <- g3s_clone(ling_imm, 'raw_report') %>%
    g3s_time(year = 2000:2002)

actions <- list(
    g3a_report_stock(agg_report, ling_imm, 'num'),
    g3a_report_stock(raw_report, ling_imm, 'num'))
# "raw_report__num" and "agg_report__num" will be available in the model report