Gadget3 spawning action

Add spawning to a g3 model

g3a_spawn_recriutment_ricker(mu, lambda)

g3a_spawn_length_straightline(alpha, beta)

g3a_spawn_length_exponential(alpha, l50)

g3a_spawn(stock, recruitment_f, proportion_f = 1, mortality_f = 0, 
          weightloss_f = 0, output_stocks = list(), 
          output_ratios = rep(1/length(output_stocks), times = length(output_stocks)), 
          mean_f, stddev_f, alpha_f, beta_f, run_f = ~TRUE, run_at = 6, 
          recruit_at = 8)

Arguments

mu
Substituted into g3a_spawn_recriutment_ricker formula, see below.

lambda
Substituted into g3a_spawn_recriutment_ricker formula, see below.

alpha
Substituted into g3a_spawn_length_* formula, see below.

beta
Substituted into g3a_spawn_length_straightline formula, see below.

l50
Substituted into g3a_spawn_length_exponential formula, see below.

stock
The mature g3_stock that will spawn in this action.

recruitment_f
A list of formula generated by one of the g3a_spawn_recruitment_* functions, containing

s

Formula run for each subset of stock

r

Final formula for calculating number of recruits for spawning action

proportion_f
formula generated by one of the g3a_spawn_length_* functions, describing the proportion of stock that will spawn at this timestep.

mortality_f
formula generated by one of the g3a_spawn_length_* functions, describing the proportion of spawning stock that will die during spawning.

weightloss_f
formula generated by one of the g3a_spawn_length_* functions, describing the overall weight loss during spawning.

output_stocks
List of g3_stocks that will be spawned into.

output_ratios
Vector of proportions for how to distribute into output_stocks, summing to 1, default evenly spread.

mean_f
formula substituted into stock structure calculations, see g3a_renewal_normalparam for details.

stddev_f
formula substituted into stock structure calculations, see g3a_renewal_normalparam for details.

alpha_f
formula substituted into stock structure calculations, see g3a_renewal_normalparam for details.

beta_f
formula substituted into stock structure calculations, see g3a_renewal_normalparam for details.

run_f
formula specifying a condition for running this action, default always runs.

run_at
Integer order that spawning actions will be run within model, see "Ordering of actions" in vignette("model_structure").

recruit_at
Integer order that recruitment from spawning will be run within model, see "Ordering of actions" in vignette("model_structure").

Details

To restrict spawning to a particular step in a year, or a particular area, use run_f. For example:

cur_step == 1

Spawning will happen on first step of every year

cur_step == 1 && cur_year >= 1990

Spawning will happen on first step of every year after 1990

cur_step == 2 && area = 1

Spawning will happen on second step of every year, in the first area

Value

g3a_spawn_recriutment_ricker

A pair of formula objects: $$ S = N_{al} p W_{al} $$ $$ R = \mu S e^{-\lambda S} $$

\(N_{al}\)

Number of parent stock

\(W_{al}\)

Weight of parent stock

\(p\)

Proportion of parent stock spawning, from proportion_f

μ

Argument provided to function

λ

Argument provided to function

g3a_spawn_length_straightline

A formula object: $$ \alpha l + \beta $$

\(l\)

Vector of stock midlengths

\(\alpha\)

alpha argument

\(\beta\)

beta argument

g3a_spawn_length_exponential

A formula object: $$ \frac{1}{1 + e^{ \alpha (l - l_{50})}} $$

\(l\)

Vector of stock midlengths

\(\alpha\)

alpha argument

\(l_{50}\)

l50 argument

g3a_spawn

An action (i.e. list of formula objects) that will, for the given stock...

Use proportion_f to calculate the total parent stock that will spawn

Use recruitment_f to derive the total newly spawned stock

Apply weightloss_f and mortality_f to the parent stock

... then, at recruitment stage ...

Recruit evenly into output_stocks, using mean_f, stddev_f, alpha_f, beta_f as-per g3a_renewal_normalparam

See also

Examples

library(magrittr)
ling_imm <- g3_stock('ling_imm', seq(20, 156, 4)) %>% g3s_age(3, 10)
ling_mat <- g3_stock('ling_mat', seq(20, 156, 4)) %>% g3s_age(3, 10)

spawn_action <- g3a_spawn(
    # Spawn from ling_mat
    ling_mat,
    # Use Ricker Recruitment Function to calculate # of recruits from total biomass
    recruitment_f = g3a_spawn_recriutment_ricker(
        ~g3_param("ricker.mu"),
        ~g3_param("ricker.lambda")),
    # Proportion of ling_mat spawning exponential relationship based on length
    proportion_f = g3a_spawn_length_exponential(
        alpha = ~g3_param("spawn.prop.alpha"), l50 = ~g3_param("spawn.prop.l50")),
    # Proportion of ling_mat dying during spawning linear relationship to length
    mortality_f = g3a_spawn_length_straightline(
        alpha = ~g3_param("spawn.mort.alpha"), beta = ~g3_param("spawn.mort.beta")),
    # Weightloss of ling_mat during spawning a constant
    weightloss_f = ~g3_param("spawn.weightloss"),
    # Spawn into ling_imm
    output_stocks = list(ling_imm),
    # Spawned stock structure, as-per g3a_renewal_normalparam()
    mean_f = 50,
    stddev_f = 0.9,
    alpha_f = 1,
    beta_f = 1,
    # Spawning should happen on the first step of every year
    run_f = ~cur_step==1)