Gadget3 timekeeping actions

Add timekeeping to a g3 model

g3a_time(start_year, end_year, steps = as.array(c(12)),
         project_years = 0L, run_at = 0)

Arguments

start_year
Year model run will start.

end_year
After this year, model run will stop.

steps
Either an MFDB time grouping, e.g. mfdb::mfdb_timestep_quarterly, or a vector of step lengths which should should sum to 12, for example, c(3,3,3,3) for quarterly steps within a year.

project_years
Number of years to continue running after the "end" of the model.

run_at
Integer order that actions will be run within model, see "Ordering of actions" in vignette("model_structure").

Details

The actions will define the following variables in your model:

cur_time

Current iteration of model, starts at 0 and increments until finished

cur_step

Current step within individual year

cur_step_size

Proportion of year this step contains, e.g. quarterly = 3/12

cur_year

Current year

cur_step_final

TRUE iff this is the final step of the year

cur_year_projection

TRUE iff we are currently projecting past end_year

total_steps

Total # of iterations (including projection) before model stops

total_years

Total # of years (including projection) before model stops

Value

g3a_time

An action (i.e. list of formula objects) that will...

Define cur_* variables listed above

If we've reached the end of the model, return nll

Examples

# Run model 2000..2004, in quarterly steps
time_action <- g3a_time(
    start_year = 2000,
    end_year = 2004,
    c(3, 3, 3, 3))