Gadget3 likelihood understocking action

Add rates of understocking in a g3 model to nll

g3l_understocking(prey_stocks, power_f = ~2, nll_breakdown = FALSE, weight = 1,
    run_at = 10)

Arguments

prey_stocks

A list of g3_stock objects to collect catch data for

power_f

A formula representing power coefficient \(p\) to use.

nll_breakdown

Should the nll report be broken down by time? TRUE / FALSE

weight

Weighting applied to this likelihood component.

run_at

Integer order that actions will be run within model, see "Ordering of actions" in vignette("model_structure").

Details

The model report will contain nll_understocking__wgt, the results of the formula below. If nll_breakdown is TRUE, this will be an array with one entry per timestep.

Value

g3l_distribution_understocking

An action (i.e. list of formula objects) that will...

Sum the total biomass adjustment due to overstocking for each prey according to the formula $$ \ell = \sum_{\it time}\sum_{\it areas} \Big(\sum_{\it prey\_stocks} U_{trs} \Big)^p $$ Where \( p \) is the power coefficient from power_f, \( U_{trs} \) is the total biomass adjustment to predator consumtion due to overconsumtion.

Examples

library(magrittr)
ling_imm <- g3_stock('ling_imm', seq(20, 156, 4)) %>% g3s_age(3, 10)
ling_mat <- g3_stock('ling_mat', seq(20, 156, 4)) %>% g3s_age(5, 15)
lln <- g3_fleet('lln')

likelihood_actions <- list(
  g3l_understocking(list(ling_imm, ling_mat)))