Gadget3 actions into R code

Convert g3 actions into an R function that can then be executed

g3_to_r(actions, trace = FALSE, strict = FALSE)

Arguments

actions

A list of actions (i.e. list of formula objects), as produced by g3a_* functions.

trace

If TRUE, turn all comments into print statements.

strict

If TRUE, enable extra sanity checking in actions. Any invalid conditions (e.g. more/less fish after growth) will result in a warning.

Value

A function that takes a params variable, that defines all g3_params required by the model. The following attributes will be set:

actions

The original actions list given to the function

parameter_template

A list of all parameters expected by the model, to fill in

Use e.g. attr(fn, 'parameter_template') to retrieve them.

Invariant model data will be stored as a closure, i.e. in environment(fn). This can be fetched with environment(fn)$cdist_sumofsquares_ldist_gil_obs__num.

The function will return nll produced by the model. You can also use attributes(nll) to get any report variables from the model.

Examples

library(magrittr)
ling_imm <- g3_stock('ling_imm', seq(20, 156, 4)) %>% g3s_age(3, 10)

initialconditions_action <- g3a_initialconditions_normalparam(
    ling_imm,
    # NB: area & age factor together (gadget2 just multiplied them)
    factor_f = ~g3_param("lingimm.init.scalar") *
        exp(-1 * (g3_param_vector("lingimm.M")[[age - 3 + 1]] + g3_param("ling.init.F")) * age) *
        g3_param_vector("lingimm.init")[[age - 3 + 1]],
    mean_f = ~g3_param("ling.Linf") * (1 - exp(-1 *
        (0.001 * g3_param("ling.k")) *
        (age - (1 + log(1 - g3_param("ling.recl") / g3_param("ling.Linf")) /
            (0.001 * g3_param("ling.k")))))),
    stddev_f = ~g3_param_vector("ling.init.sd")[[age]],
    alpha_f = ~g3_param("lingimm.walpha"),
    beta_f = ~g3_param("lingimm.wbeta"))

# Timekeeping action
time_action <- g3a_time(
    start_year = 2000,
    end_year = 2004,
    c(3, 3, 3, 3))

# Generate a model from the above 2 actions
# NB: Obviously in reality we'd need more actions
fn <- g3_to_r(list(initialconditions_action, time_action))

if (FALSE) # Edit the resulting function
fn <- edit(fn)

# Run the model for provided parameters
nll <- fn(list(
    project_years = 0,
    ling.init.F = 0.4,
    ling.Linf = 160,
    ling.k = 90,
    ling.recl = 12,
    ling.init.sd = rep(50.527220, 21),
    lingimm.init = rep(1, 15 - 5 + 1),
    lingimm.init.scalar = 200,
    lingimm.walpha = 2.27567436711055e-06,
    lingimm.wbeta = 3.20200445996187,
    lingimm.M = rep(0.15,3,10)))

# Get the report from the last model run
report <- attributes(nll)

# Fetch a value from the model data
environment(fn)$ling_imm__midlen
#>   20:24   24:28   28:32   32:36   36:40   40:44   44:48   48:52   52:56   56:60 
#>      22      26      30      34      38      42      46      50      54      58 
#>   60:64   64:68   68:72   72:76   76:80   80:84   84:88   88:92   92:96  96:100 
#>      62      66      70      74      78      82      86      90      94      98 
#> 100:104 104:108 108:112 112:116 116:120 120:124 124:128 128:132 132:136 136:140 
#>     102     106     110     114     118     122     126     130     134     138 
#> 140:144 144:148 148:152 152:156 156:Inf 
#>     142     146     150     154     158 
stopifnot(dim(environment(fn)$ling_imm__midlen) == 35)