Gadget3 actions into TMB code

Turn g3 actions into CPP code that can be compiled using TMB

g3_to_tmb(actions, trace = FALSE, strict = FALSE)

g3_tmb_adfun(cpp_code, parameters = attr(cpp_code, "parameter_template"),
    compile_flags = c("-O3", "-flto", "-march=native"), work_dir = tempdir(),
    output_script = FALSE, ...)

g3_tmb_par(parameters)

g3_tmb_lower(parameters)

g3_tmb_upper(parameters)

Arguments

actions
A list of actions (i.e. list of formula objects), as produced by g3a_* functions.

trace
If TRUE, turn all comments into print statements.

strict
If TRUE, enable extra sanity checking in actions.

cpp_code
cpp_code as produced by g3_to_tmb.

parameters
Parameter table as produced by attr(g3_to_tmb(...), 'parameter_template'), modified to provide initial conditions, etc.

compile_flags
List of extra flags to compile with, use e.g. "-g" to enable debugging output.

work_dir
Directory to write and compile .cpp files in. Defaults to R's current temporary directory

output_script
If TRUE, create a temporary R script that runs MakeADFun, and return the location. This can then be directly used with gdbsource or callr::rscript.

...
Any other options handed directly to MakeADFun

Details

g3_tmb_adfun

g3_tmb_adfun will do both the compile and MakeADFun steps of making a model. If the code is identical to an already-loaded model then it won't be recompiled, so repeated calls to g3_tmb_adfun to change parameters are fast.

If MakeADFun is crashing your R session, then you can use output_script to run in a separate R session. Use this with gdbsource to debug your model.

Value

g3_to_tmb

A string of C++ code that can be used as an input to g3_tmb_adfun.

g3_tmb_adfun

An ADFun as produced by TMB's MakeADFun, or location of temporary script if output_script is TRUE

g3_tmb_par

Values extracted from parameters table converted into a vector of values for obj$fn(par) or nlminb

g3_tmb_lower

Lower bounds extracted from parameters table converted into a vector of values for nlminb

g3_tmb_upper

Lower bounds extracted from parameters table converted into a vector of values for nlminb

Examples

library(magrittr)
ling_imm <- g3_stock('ling_imm', seq(20, 156, 4)) %>% g3s_age(3, 10)

initialconditions_action <- g3a_initialconditions_normalparam(
    ling_imm,
    # NB: area & age factor together (gadget2 just multiplied them)
    factor_f = ~g3_param("lingimm.init.scalar") *
        exp(-1 * (g3_param_vector("lingimm.M")[[age - 3 + 1]] + g3_param("ling.init.F")) * age) *
        g3_param_vector("lingimm.init")[[age - 3 + 1]],
    mean_f = ~g3_param("ling.Linf") * (1 - exp(-1 *
        (0.001 * g3_param("ling.k")) *
        (age - (1 + log(1 - g3_param("ling.recl") / g3_param("ling.Linf")) /
            (0.001 * g3_param("ling.k")))))),
    stddev_f = ~g3_param_vector("ling.init.sd")[[age]],
    alpha_f = ~g3_param("lingimm.walpha"),
    beta_f = ~g3_param("lingimm.wbeta"))

# Timekeeping action
time_action <- g3a_time(
    start_year = 2000,
    end_year = 2004,
    c(3, 3, 3, 3))

# Add our own custom action that will be added to model code
# (This is mostly to avoid problems in TMB)
custom_action <- list(~{
    nll <- g3_param('ling.k')
})

# Generate a model from the above 2 actions
# NB: Obviously in reality we'd need more actions
cpp <- g3_to_tmb(list(initialconditions_action, custom_action, time_action))

if (FALSE) # Edit the resulting code
cpp <- edit(cpp)

# Set initial conditions for parameters
tmb_param <- attr(cpp, 'parameter_template')
tmb_param$value <- I(list(
    ling.init.F = 0.4,
    ling.Linf = 160,
    ling.k = 90,
    ling.recl = 12,
    ling.init.sd = rep(50.527220, 21),
    lingimm.init = rep(1, 15 - 5 + 1),
    lingimm.init.scalar = 200,
    lingimm.walpha = 2.27567436711055e-06,
    lingimm.wbeta = 3.20200445996187,
    lingimm.M = rep(0.15,3,10))[rownames(tmb_param)])

# Compile to a TMB ADFun
tmb <- g3_tmb_adfun(cpp, tmb_param)

#> Note: Using Makevars in /home/lentinj/.R/Makevars 

if (FALSE) # Build the model in an isolated R session w/debugger
 writeLines(TMB::gdbsource(g3_tmb_adfun(cpp, compile_flags = "-g", output_script = TRUE)))

if (FALSE) # Perform a single run, using values in table
result <- tmb$fn(g3_tmb_par(tmb_param))

# perform optimisation based on parameter table
fit <- nlminb(tmb$par, tmb$fn, tmb$gr,
    upper = g3_tmb_upper(tmb_param),
    lower = g3_tmb_lower(tmb_param))

#> outer mgc:  0 

# Go back to a list of parameters, suitable for the R version
param_list <- relist(fit$par, unclass(tmb_param$value))