G3 stock_* transformation functions

Additional meta-functions to help manage writing stock-handling actions.

Details

Whilst used as functions, these functions alter the code output of the model, rather than appearing directly. They make it easier to write code that can handle stocks generically without having to worry about renaming and stock dimensions.

The transformation itself is managed by g3_step(), which all actions use.

debug_label

Add a comment to the code to act as a label for that step, when producing an outline of the model. There shouldn't be more than one debug_label call in a step.

Models compiled with trace = TRUE will print the resultant string to stdout.

Arguments

Any number of character strings, or g3_stock variables. The latter will be replaced with the final name.

Examples

> stock <- g3_stock('halibut', 1:10) |> g3s_age(1,10)
> prey_stock <- g3_stock('herring', 1:3) |> g3s_age(1,3)
>  gadget3:::g3_step(~debug_trace("Zero ", stock, "-", prey_stock, " biomass-consuming counter"))
~debug_trace("Zero halibut-herring biomass-consuming counter")

debug_trace

Identical to debug_label, but not considered a "label", just a code comment, so any number of calls can be added.

stock_assert

stock_assert(expression, message, message/stock-var, ...)

Assert that expression is true, if not abort with a message.

Examples

> stock <- g3_stock('halibut', 1:10) |> g3s_age(1,10)
> gadget3:::g3_step(~stock_assert(stock_with(stock, all(is.finite(stock__num))), stock, "__num became NaN/Inf"))
~assert_msg(~all(is.finite(halibut__num)), "halibut__num became NaN/Inf")

stock_reshape

stock_reshape(dest_stock, expression)

Output expression with it's length structure reshaped to match dest_stock. The source stock is considered to be the first one found in expression

How this is achieved depends on the difference. If the source and destination match then this is a no-op. Otherwise a transformation matrix is generated and included into the model.

Examples

stock_reshape(output_stock, stock_ss(input_stock__transitioning_num))

stock_ss

stock_ss(stock_var, [ dimname = override, dimname = override, ... ])

Subsets stock_var to get a length vector for the current iteration of stock_iterate(). If dimnames are supplied, then these dimensions overwritten with the suppled override, which could be a missing value, to preserve a dimension. See final example.

Examples

> stock <- g3_stock('halibut', 1:10) |> g3s_age(1,10)
> stock__num <- gadget3:::stock_instance(stock)
> gadget3:::g3_step(~stock_iterate(stock, { x <- x + stock_ss(stock__num) }))
~for (age in seq(halibut__minage, halibut__maxage, by = 1)) g3_with(
  halibut__age_idx := g3_idx(age - halibut__minage + 1L), {
    x <- x + stock__num[, halibut__age_idx]
  })

> stock <- g3_stock('halibut', 1:10) |> g3s_age(1,10) |> g3s_livesonareas(1)
> stock__num <- gadget3:::stock_instance(stock)
> gadget3:::g3_step(~stock_ss(stock__num, area = 5)),
~stock__num[, stock__age_idx, 5]
> gadget3:::g3_step(~stock_ss(stock__num, age = i + 1))
~stock__num[, i + 1, stock__area_idx]
> gadget3:::g3_step(~stock_ss(stock__num, area = , age = j))
~stock__num[, j, ]

stock_inv

stock_ss(stock_var, [ dimname, dimname, ... ])

like stock_ss(), but subset only the mentioned dimnames.

Examples

> stock <- g3_stock('halibut', 1:10) |> g3s_age(1,10) |> g3s_livesonareas(1)
> gadget3:::g3_step(~gadget3:::g3_step(~stock_ssinv(stock, 'age')))
~gadget3:::g3_step(~stock[, stock__age_idx, ])
> gadget3:::g3_step(~gadget3:::g3_step(~stock_ssinv(stock, 'area')))
~gadget3:::g3_step(~stock[, , stock__area_idx])

stock_switch

stock_ss(stock, stock_name1 = expr, stock_name2 = expr, ... [ default ])

Switch based on name of stock, returning the relevant expr or default. If no default supplied, then an unknown stock is an error.

expr is implicitly wrapped with stock_with(stock, ...), so any references to the stock variable will work. If only default is provided, then this is identical to calling stock_with.

Examples

> stock <- g3_stock('halibut', 1:10) ; fleet_stock <- g3_fleet('igfs')
> gadget3:::g3_step(~stock_switch(stock, halibut = 2, herring = 3, -1))
~2
> gadget3:::g3_step(~stock_switch(fleet_stock, halibut = 2, herring = 3, -1))
~-1
> gadget3:::g3_step(~stock_switch(stock, halibut = stock__midlen, -1))
~halibut__midlen

stock_with

stock_with(stock, expr)

Replaced with expr but with all stock variables of stock renamed with their final name. This is generally needed when not iterating over a stock, but e.g. zeroing or summing the whole thing.

Examples

> stock <- g3_stock('halibut', 1:10)
> gadget3:::g3_step(~stock_with(stock, sum(stock__num)))
~sum(halibut__num)

stock_iterate

stock_iterate(stock, expr)

Wrap expr with the code to iterate over each length vector in stock, accessed using stock_ss(stock)

Current values for each dimension will be available as variables, e.g. area, age, and can be used in formulae.

Examples

> stock <- g3_stock('halibut', 1:10) |> g3s_age(1,10)
> gadget3:::g3_step(~stock_iterate(stock, x <- x + stock_ss(stock__num)))
~for (age in seq(halibut__minage, halibut__maxage, by = 1)) g3_with(`:=`(halibut__age_idx,
    g3_idx(age - halibut__minage + 1L)), x <- x + halibut__num[, halibut__age_idx])

stock_intersect

stock_intersect(stock, expr)

Wrap expr with the code to intersect all dimensions with the dimensions of an outer stock_iterate().

Examples

> stock <- g3_stock('halibut', 1:10) |> g3s_age(1,10)
> prey_stock <- g3_stock('herring', 1:3) |> g3s_age(1,3)
> gadget3:::g3_step(~stock_iterate(stock, stock_intersect(prey_stock, x <- x + stock_ss(stock__num) + stock_ss(prey_stock__num) )))
~for (age in seq(halibut__minage, halibut__maxage, by = 1)) if (age >=
    herring__minage && age <= herring__maxage) g3_with(
      halibut__age_idx := g3_idx(age - halibut__minage + 1L),
      herring__age_idx := g3_idx(age - herring__minage + 1L), {
        x <- x + halibut__num[, halibut__age_idx] + herring__num[, herring__age_idx]
      })

stock_interact

stock_interact(stock, expr, prefix = prefix)

Wrap expr with the code to interact with the dimensions of an outer stock_iterate(). Interact means to intersect over area, but try the combinatoral explosion of all other dimensions, i.e. what would make most sense when 2 stocks interact in a predator-prey relationship.

Additional variables will be prefixed with prefix.

Examples

> stock <- g3_stock('halibut', 1:10) |> g3s_age(1,10)
> prey_stock <- g3_stock('herring', 1:3) |> g3s_age(1,3)
> gadget3:::g3_step(~stock_iterate(stock, stock_interact(prey_stock, {
    x <- x + stock_ss(stock__num) + stock_ss(prey_stock__num)}, prefix = "prey" )))
~for (age in seq(halibut__minage, halibut__maxage, by = 1)) g3_with(
  halibut__age_idx := g3_idx(age - halibut__minage + 1L), {
    for (prey_age in seq(herring__minage, herring__maxage, by = 1)) g3_with(
      herring__age_idx := g3_idx(prey_age - herring__minage + 1L), {
        x <- x + halibut__num[, halibut__age_idx] + herring__num[, herring__age_idx]
      }
  }))

stock_param

stock_param(stock, name, name_part = NULL, ...)

Convert to a call to g3_param, prefixing the name with the stock name. If name_part given, only use that part of the stock name.

All other args passed through to g3_param.

Examples

> stock <- g3_stock(c(species = 'halibut', 'imm'), 1:10) |> g3s_age(1,10)
> gadget3:::g3_step(~stock_param(stock, 'K'))
~g3_param("halibut_imm.K")
> gadget3:::g3_step(~stock_param(stock, 'K', name_part = 'species'))
~g3_param("halibut.K")

stock_param_table

stock_param_table(stock, name, name_part = NULL, table_defn, ...)

Convert to a call to g3_param_table, prefixing the name with the stock name. If name_part given, only use that part of the stock name.

The table_defn will have any instances of stock replaced using stock_with().

All other args passed through to g3_param_table.

Examples

> stock <- g3_stock(c(species = 'halibut', 'imm'), 1:10) |> g3s_age(1,10)
> gadget3:::g3_step(~stock_param_table(stock, 'K', expand.grid(year = 2000:2004)))
~g3_param_table("halibut_imm.K", expand.grid(year = 2000:2004))
> gadget3:::g3_step(~stock_param_table(stock, 'K', expand.grid(age = seq(stock__minage, stock__maxage))))
~g3_param_table("halibut_imm.K", expand.grid(age = seq(halibut_imm__minage,
    halibut_imm__maxage)))